changed get_distance function return type from unsigned int to double

+++ point.c
@@ -1,14 +1,14 @@
 #include <math.h>
 #include "point.h"
 
-unsigned int get_distance(Point point_A, Point point_B)
+double get_distance(Point point_A, Point point_B)
 {
   return sqrt(pow(point_A.x - point_B.x, 2) + pow(point_A.y - point_B.y, 2));
 }
 
 void get_closest_food(Point food_points[], int points_length, Point current_location, Point *closest_food_location)
 {
-  unsigned int distance, closest_distance = MAX_VALUE;
+  double distance, closest_distance = MAX_VALUE;
 
   DO_LENGTH_TIMES
   {
